requirements.txt
sqlglot>=23.0.0
neo4j>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.0
openai>=1.0.0
anthropic>=0.30.0
certifi>=2024.0.0
sse-starlette>=2.0.0
PyYAML>=6.0.0
# Cache
redis>=5.0.0
# Observability
prometheus-client>=0.20.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
# Test dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
anyio[trio]>=4.0.0
httpx>=0.27.0
# OWL Reasoner
rdflib>=7.0.0
owlrl>=7.0.0
pyshacl>=0.25.0
